Updates display logic for recovered stations
Updates station map rules for displaying recovered stations.
Previously, only stations with no obs data were considered recovered,
and also ignored the contents of the "recovered" field.  Now, if a
station with observations has "recovered" set to true, or has no data
for ten or more days, it is displayed as recovered in addition to the
existing case with no observations.